requirements.txt
torch>=1.12.0
torch-geometric>=2.1.0
torchvision>=0.13.0
scikit-learn>=1.1.0
numpy>=1.21.0
pandas>=1.4.0
transformers>=4.20.0
sentence-transformers>=2.2.0
emoji>=2.0.0
tqdm>=4.64.0
python-dateutil>=2.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
networkx>=2.8.0
requests>=2.28.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
pyyaml>=6.0